requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
asyncpg==0.29.0